Prototyped a generalised simulation system
I have developed a system to allow for the modelling of the simulation using an xml file.

The system allows for creation of Sensors, Gateways and Servers along with their links between each other. When sending packets to devices outside the immediate connection, a Depth First Search will try and find a route through the device links and box the packet in a RoutingPacket to its destination.

The system allows for Devices to run through simulation files to communicate with each other.